added a folder for footer pages and footer components
