+++ react/src/components/footer/About.jsx
