+++ react/src/components/footer/PrivacyPolicy.jsx
