+++ react/src/components/footer/Terms.jsx
