--- react/src/pages/AboutUs.jsx
@@ -1,13 +1,0 @@
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import About from "../components/About";
-
-export default function AboutUs() {
-  return (
-    <>
-      <Header></Header>
-      <About></About>
-      <Footer></Footer>
-    </>
-  );
-}
