--- react/src/pages/CareersPage.jsx
@@ -1,13 +1,0 @@
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import Careers from "../components/Careers";
-
-export default function CareersPage() {
-  return (
-    <>
-      <Header></Header>
-      <Careers></Careers>
-      <Footer></Footer>
-    </>
-  );
-}
