+++ react/src/pages/CompanyOverview.jsx
@@ -1,4 +1,4 @@
-import Footer from "../components/Footer";
+import Footer from "../components/footer/Footer";
 import BasicTabs from "../components/Tabs";
 import { Box, Typography } from "@mui/material";
 import { useNavigate } from "react-router-dom";
