--- react/src/pages/ContactUs.js
@@ -1,13 +1,0 @@
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import Contact from "../components/Contact";
-
-export default function AboutUs() {
-  return (
-    <>
-      <Header></Header>
-      <Contact></Contact>
-      <Footer></Footer>
-    </>
-  );
-}
