+++ react/src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import Banner from "../components/Banner";
 import CarouselSlide from "../components/CarouselCards";
 //import Header from "../components/Header";
-import Footer from "../components/Footer";
+import Footer from "../components/footer/Footer";
 
 export default function Homepage({ onSelect, setCompany }) {
   return (
