+++ react/src/pages/Other/AboutUs.jsx
@@ -1,0 +1,13 @@
+import Header from "../../components/Header";
+import Footer from "../../components/footer/Footer";
+import About from "../../components/footer/About";
+
+export default function AboutUs() {
+  return (
+    <>
+      <Header></Header>
+      <About></About>
+      <Footer></Footer>
+    </>
+  );
+}
