+++ react/src/pages/Other/CareersPage.jsx
@@ -1,0 +1,13 @@
+import Footer from "../../components/footer/Footer";
+import Header from "../../components/Header";
+import Careers from "../../components/footer/Careers";
+
+export default function CareersPage() {
+  return (
+    <>
+      <Header></Header>
+      <Careers></Careers>
+      <Footer></Footer>
+    </>
+  );
+}
