+++ react/src/pages/Other/ContactUs.js
@@ -1,0 +1,13 @@
+import Header from "../../components/Header";
+import Footer from "../../components/footer/Footer";
+import Contact from "../../components/footer/Contact";
+
+export default function AboutUs() {
+  return (
+    <>
+      <Header></Header>
+      <Contact></Contact>
+      <Footer></Footer>
+    </>
+  );
+}
