+++ react/src/pages/Other/Privacy.jsx
@@ -1,0 +1,13 @@
+import Header from "../../components/Header";
+import Footer from "../../components/footer/Footer";
+import PrivacyPolicy from "../../components/footer/PrivacyPolicy";
+
+export default function Privacy() {
+  return (
+    <>
+      <Header></Header>
+      <PrivacyPolicy></PrivacyPolicy>
+      <Footer></Footer>
+    </>
+  );
+}
