+++ react/src/pages/Other/Team.jsx
@@ -1,0 +1,13 @@
+import Header from "../../components/Header";
+import Footer from "../../components/footer/Footer";
+import Teamjob from "../../components/footer/Teamjob";
+
+export default function Team() {
+  return (
+    <>
+      <Header></Header>
+      <Teamjob></Teamjob>
+      <Footer></Footer>
+    </>
+  );
+}
