+++ react/src/pages/Other/TermsofUse.jsx
@@ -1,0 +1,13 @@
+import Header from "../../components/Header";
+import Footer from "../../components/footer/Footer";
+import Terms from "../../components/footer/Terms";
+
+export default function TermsofUse() {
+  return (
+    <>
+      <Header></Header>
+      <Terms></Terms>
+      <Footer></Footer>
+    </>
+  );
+}
