--- react/src/pages/Privacy.jsx
@@ -1,13 +1,0 @@
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import PrivacyPolicy from "../components/PrivacyPolicy";
-
-export default function Privacy() {
-  return (
-    <>
-      <Header></Header>
-      <PrivacyPolicy></PrivacyPolicy>
-      <Footer></Footer>
-    </>
-  );
-}
