--- react/src/pages/Team.jsx
@@ -1,13 +1,0 @@
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import Teamjob from "../components/Teamjob";
-
-export default function Team() {
-  return (
-    <>
-      <Header></Header>
-      <Teamjob></Teamjob>
-      <Footer></Footer>
-    </>
-  );
-}
