--- react/src/pages/TermsofUse.jsx
@@ -1,13 +1,0 @@
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import Terms from "../components/Terms";
-
-export default function TermsofUse() {
-  return (
-    <>
-      <Header></Header>
-      <Terms></Terms>
-      <Footer></Footer>
-    </>
-  );
-}
